* Delta time (dt) now calculated properly using PerfTimer.   World Update now called in OnIdle() - works much better than before   Warning: Don't call gameTime Update() more than once per frame

+++ 2DPhysicsDemo/World.h
@@ -35,7 +35,7 @@ public:
 	World();
 	~World();
 
-	void Update(f32 dt);
+	void Update(f64 dt);
 	void Draw();
 	void Load();
 	void Unload();
